DBG - reproducer for https://issues.chromium.org/issues/400486352
After checking out this branch, run this in a terminal:

    TEST_SHOW_BROWSER=1 ./test/verify/chromedriver-websocket-traffic

A Chromium browser window should open with a "BIDI tab". Use it to
manually log into any machine that runs Cockpit.  And observe the
"BIDI tab": It logs all of Cockpit's own WebSocket traffic.

Maybe upload a file using the "File browser" of Cockpit. The content
of that file will appear in base64 encoded form in the "BIDI" tab.

You can install Cockpit on many Linux operating systems from
distribution packages, see

    https://cockpit-project.org/running.html

Once installed and enabled/started, you can access it on port 9090 and
log in as any user of the machine.